rename 'parent' to 'user' for dashboard and components
